create in-app notifiation contents component
